+++ app/Http/Controllers/Backend/ProductoController.php
@@ -45,8 +45,9 @@ class ProductoController extends AppBaseController
      */
     public function create()
     {
-        $categorias = Categoria::all();
-        return view('backend.productos.create')->with('categorias',$categorias);
+        $categorias = Categoria::all()->lists('nombre', 'id');
+        return view('backend.productos.create')
+                ->with('categorias',$categorias);
     }
 
     /**
@@ -64,7 +65,7 @@ class ProductoController extends AppBaseController
         Storage::disk('public')->put($newFilename, file_get_contents($file));       
         $input['imagen'] = /*storage_path('app/public'). DIRECTORY_SEPARATOR .*/$newFilename;
         $producto = $this->productoRepository->create($input);
-        Flash::success('Producto saved successfully.');
+        Flash::success('Producto guardado exitosamente');
         return redirect(route('backend.productos.index'));
     }
 
@@ -80,12 +81,13 @@ class ProductoController extends AppBaseController
         $producto = $this->productoRepository->findWithoutFail($id);
 
         if (empty($producto)) {
-            Flash::error('Producto not found');
+            Flash::error('Producto no encontrado');
 
             return redirect(route('backend.productos.index'));
         }
 
-        return view('backend.productos.show')->with('producto', $producto);
+        return view('backend.productos.show')
+                ->with('producto', $producto);
     }
 
     /**
@@ -100,13 +102,16 @@ class ProductoController extends AppBaseController
         $producto = $this->productoRepository->findWithoutFail($id);
 
         if (empty($producto)) {
-            Flash::error('Producto not found');
+            Flash::error('Producto no encontrado');
 
             return redirect(route('backend.productos.index'));
         }
-        $categorias = Categoria::all();
-
-        return view('backend.productos.edit')->with('producto', $producto)->with('categorias',$categorias);
+        $categorias = Categoria::where('id','<>',$producto->categoria_id)->lists('nombre', 'id');
+        $categoria = Categoria::where('id',$producto->categoria_id)->lists('nombre', 'id');
+        $categorias = $categoria->union($categorias);
+        return view('backend.productos.edit')
+                ->with('producto', $producto)
+                ->with('categorias',$categorias);
     }
 
     /**
@@ -122,14 +127,14 @@ class ProductoController extends AppBaseController
         $producto = $this->productoRepository->findWithoutFail($id);
 
         if (empty($producto)) {
-            Flash::error('Producto not found');
+            Flash::error('Producto no encontrado');
 
             return redirect(route('backend.productos.index'));
         }
-
+        
         $producto = $this->productoRepository->update($request->all(), $id);
 
-        Flash::success('Producto updated successfully.');
+        Flash::success('Producto actualizado exitosamente');
 
         return redirect(route('backend.productos.index'));
     }
@@ -146,14 +151,19 @@ class ProductoController extends AppBaseController
         $producto = $this->productoRepository->findWithoutFail($id);
 
         if (empty($producto)) {
-            Flash::error('Producto not found');
+            Flash::error('Producto no encontrado');
 
             return redirect(route('backend.productos.index'));
         }
+        $articulos = $producto->articulos()->get();
+        if (!empty($articulos)){
+            Flash::error('No es posible eliminar el Producto dado que tiene Articulos asociados');
 
+            return redirect(route('backend.categorias.index'));
+        }
         $this->productoRepository->delete($id);
 
-        Flash::success('Producto deleted successfully.');
+        Flash::success('Producto borrado exitosamente');
 
         return redirect(route('backend.productos.index'));
     }
